added more dtos
