+++ Backend/tutofinder-tutorship-service/src/main/java/com/tutofinder/tutorship/dto/AddressDto.java
@@ -1,4 +1,4 @@
-package com.tutofinder.tutorship.dtos;
+package com.tutofinder.tutorship.dto;
 
 import lombok.Data;
