+++ Backend/tutofinder-tutorship-service/src/main/java/com/tutofinder/tutorship/dto/FatherDto.java
@@ -1,20 +1,18 @@
-package com.tutofinder.tutorship.dtos;
+package com.tutofinder.tutorship.dto;
 
 import lombok.Data;
-
 import java.util.Date;
+import java.util.List;
 
 @Data
-public class TeacherDto {
+public class FatherDto {
     private Long id;
     private String firstName;
     private String lastName;
     private String dni;
     private String email;
-    private String depositNumber;
-    private Double hourlyCost;
-    private String membership;
     private AddressDto address;
+    private List<StudentDto> students;
     private Date createdDate;
     private Date lastUpdateDate;
 }
