+++ Backend/tutofinder-tutorship-service/src/main/java/com/tutofinder/tutorship/dto/StudentDto.java
@@ -1,0 +1,20 @@
+package com.tutofinder.tutorship.dto;
+
+import com.tutofinder.tutorship.util.EducationLevel;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+public class StudentDto {
+    private Long id;
+    private String firstName;
+    private String lastName;
+    private String dni;
+    private String email;
+    private EducationLevel educationLevel;
+    private Date createdDate;
+    private Date lastUpdateDate;
+}
