+++ Backend/tutofinder-tutorship-service/src/main/java/com/tutofinder/tutorship/dto/create/CreateTeacherDto.java
@@ -1,6 +1,6 @@
-package com.tutofinder.tutorship.dtos.create;
+package com.tutofinder.tutorship.dto.create;
 
-import com.tutofinder.tutorship.dtos.AddressDto;
+import com.tutofinder.tutorship.dto.AddressDto;
 
 import lombok.Data;
